refactor(app): fold internal/ui into internal/app/render
internal/ui was a three-file package holding one type — the overlay
renderer facade — under a top-level name that sounded like a layer but
was not one: it sat outside the hexagonal vocabulary (domain, ports,
adapter, app, cli, config), and the layering architecture test had to
special-case it in three places to keep the rules coherent.

The renderer now lives at internal/app/render, which is what it always
was: application-layer code driving the overlay adapter. The layering
test's special cases are deleted, so the layer rules close one more
hole, and the docs' component diagram and layer tables describe the
real shape. Pure move; no behavior change.
