+++ internal/app/modes/recursive_grid_test.go
@@ -12,12 +12,12 @@ import (
 	hintscomponent "github.com/y3owk1n/neru/internal/adapter/overlay/render/hints"
 	componentrecursivegrid "github.com/y3owk1n/neru/internal/adapter/overlay/render/recursivegrid"
 	"github.com/y3owk1n/neru/internal/app/components"
+	"github.com/y3owk1n/neru/internal/app/render"
 	"github.com/y3owk1n/neru/internal/app/services"
 	"github.com/y3owk1n/neru/internal/config"
 	"github.com/y3owk1n/neru/internal/domain"
 	"github.com/y3owk1n/neru/internal/domain/state"
 	portmocks "github.com/y3owk1n/neru/internal/ports/mocks"
-	"github.com/y3owk1n/neru/internal/ui"
 )
 
 func TestHandleRecursiveGridKey_CompleteSelectionDoesNotMoveWhenCursorFollowSelectionDisabled(
@@ -112,7 +112,7 @@ func TestResetCurrentMode_RecursiveGridPreservesHoldMode(t *testing.T) {
 			},
 			zap.NewNop(),
 		),
-		renderer: ui.NewOverlayRenderer(
+		renderer: render.NewOverlayRenderer(
 			&overlaypkg.NoOpManager{},
 			hintscomponent.StyleMode{},
 			gridcomponent.Style{},
