+++ internal/app/render/doc.go
@@ -1,6 +1,8 @@
-// Package ui implements the user interface components for the Neru application.
+// Package render is the app-layer rendering facade: it owns the
+// OverlayRenderer that translates mode state into calls on the overlay
+// manager and the shared render styles.
 //
 // This package provides the visual interface layer for Neru, managing all aspects
 // of on-screen rendering and user interaction. It serves as the bridge between
 // the application's core logic and the visual elements displayed to users.
-package ui
+package render
