+++ internal/app/render/renderer.go
@@ -1,4 +1,4 @@
-package ui
+package render
 
 import (
 	"image"
